設定画面と表示画面のファイル分割
